Split is-external tests.
