+++ client/lib/url/test/is-external.js
@@ -1,0 +1,143 @@
+/**
+ * @jest-environment jsdom
+ */
+
+/**
+ * Internal dependencies
+ */
+import isExternal from '../is-external';
+
+describe( 'isExternal', () => {
+	test( 'should return false for relative path-only url', () => {
+		const source = '/relative';
+
+		const actual = isExternal( source );
+
+		expect( actual ).toBeFalse;
+	} );
+
+	test( 'should return false for relative query-only url', () => {
+		const source = '?foo=bar';
+
+		const actual = isExternal( source );
+
+		expect( actual ).toBeFalse;
+	} );
+
+	test( 'should return true for url without http', () => {
+		const urlWithoutHttp = 'en.support.wordpress.com';
+		const actual = isExternal( urlWithoutHttp );
+		expect( actual ).toBeTrue;
+	} );
+
+	test( 'should return true for url without http and path', () => {
+		const urlWithoutHttp = 'en.support.wordpress.com/start';
+		const actual = isExternal( urlWithoutHttp );
+		expect( actual ).toBeTrue;
+	} );
+
+	test( 'should return true for url without http and // path', () => {
+		const urlWithoutHttp = 'en.support.wordpress.com//start';
+		const actual = isExternal( urlWithoutHttp );
+		expect( actual ).toBeTrue;
+	} );
+
+	test( 'should return true for just external relative // path', () => {
+		const urlWithoutHttp = '//manage';
+		const actual = isExternal( urlWithoutHttp );
+		expect( actual ).toBeTrue;
+	} );
+
+	describe( 'with global.window (test against window.location.hostname)', () => {
+		// window.location.hostname === 'example.com'
+		test( 'should return false for for internal protocol-relative url', () => {
+			const source = '//example.com';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeFalse;
+		} );
+
+		test( 'should return true for for external protocol-relative url', () => {
+			const source = '//some-other-site.com';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeTrue;
+		} );
+
+		test( 'should return false for internal url', () => {
+			const source = 'https://example.com';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeFalse;
+		} );
+
+		test( 'should return true for external url', () => {
+			const source = 'https://not-example.com';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeTrue;
+		} );
+
+		test( 'should return false for internal path without http', () => {
+			const urlWithoutHttp = 'example.com//me';
+			const actual = isExternal( urlWithoutHttp );
+			expect( actual ).toBeFalse;
+		} );
+
+		test( 'should return true for internal but legacy path with http', () => {
+			const urlWithoutHttp = 'http://example.com/manage';
+			const actual = isExternal( urlWithoutHttp );
+			expect( actual ).toBeTrue;
+		} );
+
+		test( 'should return true for internal but legacy path without http', () => {
+			const urlWithoutHttp = 'example.com/manage';
+			const actual = isExternal( urlWithoutHttp );
+			expect( actual ).toBeTrue;
+		} );
+
+		test( 'should return true for subdomains', () => {
+			const source = '//ns1.example.com';
+			const actual = isExternal( source );
+			expect( actual ).toBeTrue;
+		} );
+	} );
+
+	describe( 'without global.window (test against config hostname)', () => {
+		test( 'should return false for internal protocol-relative url', () => {
+			const source = '//calypso.localhost';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeFalse;
+		} );
+
+		test( 'should return true for external protocol-relative url', () => {
+			const source = '//some-other-site.com';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeTrue;
+		} );
+
+		test( 'should return false for internal url', () => {
+			const source = 'https://calypso.localhost/me';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeFalse;
+		} );
+
+		test( 'should return true for external url', () => {
+			const source = 'https://not.localhost/me';
+
+			const actual = isExternal( source );
+
+			expect( actual ).toBeTrue;
+		} );
+	} );
+} );
